Add make_choice file that contains list of options the user to choose from
