+++ item.rb
@@ -1,11 +1,11 @@
 require 'date'
 class Item
-  def initialize(id, genre, source, label, publish_date, archived)
-    @id = id
-    @genre = genre
-    @source = source
-    @label = label
-    @publish_date = publish_date
+  attr_reader :genre, :source, :label
+  attr_accessor :publish_date, :archived
+
+  def initialize(_id, archived: false)
+    @id = Random.rand(1..1000)
+    @publish_date = date
     @archived = archived
   end
 
@@ -15,6 +15,11 @@ class Item
 
   def move_to_archive
     @archived = true if can_be_archived?
+  end
+
+  def genre=(genre)
+    @genre = genre
+    genre.items << self unless genre.items.includes?(self)
   end
   private :can_be_archived?
 end
